requirements.txt
upstash-vector
groq
python-dotenv
requests
Flask
asyncpg